EMP detection, insult on kill, aim improvement, flag and freq commands

+++ screenbot/Config.cpp
@@ -23,7 +23,7 @@ namespace Convert {
     template<>
     float Get(const tstring& val) {
         if (val.length() == 0) return 0.0f;
-        return _tstof(val.c_str());
+        return (float)_tstof(val.c_str());
     }
 
     template<>
